Commit Inicio interaccion Bonus
Creados metodos polimorficos de Vector2D.
Creado metodo de colision bonusespecial
Siguiente: en mundo implementar acciones
