+++ src/dominio/Campo.cpp
@@ -1,5 +1,5 @@
 #include "..\..\include\dominio\Campo.h"
-#include "..\include\comun\ETSIDI.h"
+#include "ETSIDI.h"
 
 Campo::Campo()  //constructor de campo. Damos valores iniciales
 {
@@ -31,23 +31,25 @@ void Campo::Dibuja(){
 	pared_dcha.Dibuja();
 	pared_izq.Dibuja();
 	red.Dibuja();
+	
 	/*
 	//dibujo del fondo
 	glEnable(GL_TEXTURE_2D);
 
-	glBindTexture(GL_TEXTURE_2D, ETSIDI::getTexture("..\..\bin\imagenes\fondo.png").id);
+	glBindTexture(GL_TEXTURE_2D, ETSIDI::getTexture("imagenes/fondo.png").id);
 	glDisable(GL_LIGHTING);
 	glBegin(GL_POLYGON);
 	glColor3f(1, 1, 1);
 
-	glTexCoord2d(0, 1);		glVertex3f(-10, 0, -0.1);
-	glTexCoord2d(1, 1);		glVertex3f(10, 0, -0.1);
-	glTexCoord2d(1, 0);		glVertex3f(10, 15, -0.1);
-	glTexCoord2d(0, 0);		glVertex3f(-10, 15, -0.1);
+	glTexCoord2d(0, 1);		glVertex2f(-10, 0);
+	glTexCoord2d(1, 1);		glVertex2f(10, 0);
+	glTexCoord2d(1, 0);		glVertex2f(10, 15);
+	glTexCoord2d(0, 0);		glVertex2f(-10, 15);
 	glEnd();
+
+	glPopMatrix();
 
 	glEnable(GL_LIGHTING);
 	glDisable(GL_TEXTURE_2D);
-
 	*/
 }
