--- tmp/Debug/Juego.Build.CppClean.log
@@ -1,24 +1,0 @@
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\vc120.pdb
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\vc120.idb
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\vector2d.obj
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\balon.obj
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\bonus.obj
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\bonusespecial.obj
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\bonusnormal.obj
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\campo.obj
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\hombre.obj
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\interaccion.obj
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\objetomovil.obj
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\pared.obj
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\coordinadorjuego.obj
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\juego.obj
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\mundo.obj
-c:\users\aaron\desktop\2016\practicas\juego\bin\juego.ilk
-c:\users\aaron\desktop\2016\practicas\juego\bin\juego.exe
-c:\users\aaron\desktop\2016\practicas\juego\bin\juego.pdb
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\juego.tlog\cl.command.1.tlog
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\juego.tlog\cl.read.1.tlog
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\juego.tlog\cl.write.1.tlog
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\juego.tlog\link.command.1.tlog
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\juego.tlog\link.read.1.tlog
-c:\users\aaron\desktop\2016\practicas\juego\tmp\debug\juego.tlog\link.write.1.tlog
