+++ tmp/Debug/Pang.log
@@ -1,25 +1,8 @@
-﻿Compilación iniciada a las 5/29/2016 7:56:16 PM.
-     1>Proyecto "C:\Users\Aaron\Desktop\2016\Practicas\Juego\Pang.vcxproj" en el nodo 2 (Rebuild destinos).
-     1>C:\Program Files (x86)\MSBuild\Microsoft.Cpp\v4.0\V120\Microsoft.CppBuild.targets(381,5): warning MSB8028: The intermediate directory (tmp\Debug\) contains files shared from another project (Pang.vcxproj).  This can lead to incorrect clean and rebuild behavior.
+﻿Compilación iniciada a las 5/29/2016 11:13:53 PM.
+     1>Proyecto "C:\Users\Aaron\Desktop\2016\Practicas\Juego\Pang.vcxproj" en el nodo 2 (Build destinos).
      1>ClCompile:
-         C:\Program Files (x86)\Microsoft Visual Studio 12.0\VC\bin\CL.exe /c /Ilib /Isrc /ZI /nologo /W3 /WX- /Od /Oy- /D WIN32 /D _DEBUG /D _CONSOLE /D _UNICODE /D UNICODE /Gm /EHsc /RTC1 /MDd /GS /fp:precise /Zc:wchar_t /Zc:forScope /Fo"tmp\Debug\\" /Fd"tmp\Debug\vc120.pdb" /Gd /TP /analyze- /errorReport:prompt src\comun\Vector2D.cpp src\dominio\Balon.cpp src\dominio\Bonus.cpp src\dominio\BonusEspecial.cpp src\dominio\BonusNormal.cpp src\dominio\Campo.cpp src\dominio\Hombre.cpp src\dominio\Interaccion.cpp src\dominio\ObjetoMovil.cpp src\dominio\Pared.cpp src\juego\CoordinadorJuego.cpp src\juego\Juego.cpp src\juego\Mundo.cpp
+         C:\Program Files (x86)\Microsoft Visual Studio 12.0\VC\bin\CL.exe /c /Ilib /Isrc /ZI /nologo /W3 /WX- /Od /Oy- /D WIN32 /D _DEBUG /D _CONSOLE /D _UNICODE /D UNICODE /Gm /EHsc /RTC1 /MDd /GS /fp:precise /Zc:wchar_t /Zc:forScope /Fo"tmp\Debug\\" /Fd"tmp\Debug\vc120.pdb" /Gd /TP /analyze- /errorReport:prompt src\juego\Mundo.cpp
          Mundo.cpp
-         Juego.cpp
-         CoordinadorJuego.cpp
-         Pared.cpp
-         ObjetoMovil.cpp
-         Interaccion.cpp
-         Hombre.cpp
-         Campo.cpp
-         BonusNormal.cpp
-         BonusEspecial.cpp
-     1>c:\users\aaron\desktop\2016\practicas\juego\src\dominio\bonusespecial.cpp(26): warning C4244: '=' : conversión de 'float' a 'unsigned char'; posible pérdida de datos
-     1>c:\users\aaron\desktop\2016\practicas\juego\src\dominio\bonusespecial.cpp(27): warning C4244: '=' : conversión de 'float' a 'unsigned char'; posible pérdida de datos
-     1>c:\users\aaron\desktop\2016\practicas\juego\src\dominio\bonusespecial.cpp(28): warning C4244: '=' : conversión de 'float' a 'unsigned char'; posible pérdida de datos
-         Bonus.cpp
-         Balon.cpp
-         Vector2D.cpp
-         Generando código...
        Link:
          C:\Program Files (x86)\Microsoft Visual Studio 12.0\VC\bin\link.exe /ERRORREPORT:PROMPT /OUT:"C:\Users\Aaron\Desktop\2016\Practicas\Juego\bin\Juego.exe" /INCREMENTAL /NOLOGO /LIBPATH:lib kernel32.lib user32.lib gdi32.lib winspool.lib comdlg32.lib advapi32.lib shell32.lib ole32.lib oleaut32.lib uuid.lib odbc32.lib odbccp32.lib /MANIFEST /MANIFESTUAC:"level='asInvoker' uiAccess='false'" /manifest:embed /DEBUG /PDB:"C:\Users\Aaron\Desktop\2016\Practicas\Juego\bin\Juego.pdb" /SUBSYSTEM:CONSOLE /TLBID:1 /DYNAMICBASE /NXCOMPAT /IMPLIB:"C:\Users\Aaron\Desktop\2016\Practicas\Juego\bin\Juego.lib" /MACHINE:X86 tmp\Debug\Vector2D.obj
          tmp\Debug\Balon.obj
@@ -35,8 +18,8 @@
          tmp\Debug\Juego.obj
          tmp\Debug\Mundo.obj
          Pang.vcxproj -> C:\Users\Aaron\Desktop\2016\Practicas\Juego\bin\Juego.exe
-     1>Compilación del proyecto terminada "C:\Users\Aaron\Desktop\2016\Practicas\Juego\Pang.vcxproj" (Rebuild destinos).
+     1>Compilación del proyecto terminada "C:\Users\Aaron\Desktop\2016\Practicas\Juego\Pang.vcxproj" (Build destinos).
 
 Compilación correcta.
 
-Tiempo transcurrido 00:00:02.40
+Tiempo transcurrido 00:00:01.12
