--- tmp/Debug/Pang.tlog/Pang.lastbuildstate
@@ -1,2 +1,0 @@
-#TargetFrameworkVersion=v4.0:PlatformToolSet=v120:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit
-Debug|Win32|C:\Users\Aaron\Desktop\2016\Practicas\Practica4\Tema4_inicial_2012\Pang\|
